requirements.txt
lmdb==1.3.0
opencv-python==4.6.0.66
tqdm==4.64.1
numpy==1.23.4
Pillow==9.2.0

